Reorganize time and time utils to date_time lib
